+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Added:
 - Add data source for `chronosphere_collection`
 - Add support for `drop_nan_value` in `chronosphere_drop_rule` resource
 - Add `gcp_metrics_integration` resource
+- Add new resource type, `dataset`, with the first telemetry type supported being tracing.
 
 Deprecated:
 - Remove unsupported `bucket_id` in `chronosphere_notification_policy`